Accel now use int with 0.1° precision, but doesn't work

+++ accel.h
@@ -4,4 +4,4 @@ void rcv_axis();
 #define BUF_LEN 9
 extern char axes[BUF_LEN][3];
 extern unsigned char buf_index;
-extern code float lut_arccos[32];
+extern code int lut_arccos[32];
